requirements.txt
Django==1.10.4
hashids==1.2.0
isort==4.2.5
MySQL-python==1.2.5
pillow==3.4.2
#pymssql==2.1.3
python-memcached==1.58
python-slugify==1.2.1
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.20
django-simple-captcha==0.5.3
django-ckeditor==5.2.1
django-el-pagination==3.0.1
django-debug-toolbar==1.7
xlwt==1.2.0
pytz==2017.2
